Move API caller to its own file
